+++ h2gis-functions/src/test/java/org/h2gis/functions/io/fgb/FGBImportExportTest.java
@@ -361,9 +361,29 @@ public class FGBImportExportTest {
         try (Statement stat = connection.createStatement()) {
             stat.execute("DROP TABLE IF EXISTS TABLE_POINTS");
             stat.execute("create table TABLE_POINTS(id int, the_geom GEOMETRY(POINT, 4326))");
-            stat.execute("CALL FGBWrite('/tmp/empty_file.fgb', 'TABLE_POINTS', true);");
+            stat.execute("CALL FGBWrite('target/empty_file.fgb', 'TABLE_POINTS', true);");
             stat.execute("DROP TABLE IF EXISTS TABLE_POINTS");
-            stat.execute("CALL FGBRead('/tmp/empty_file.fgb', 'TABLE_POINTS', true);");
+            stat.execute("CALL FGBRead('target/empty_file.fgb', 'TABLE_POINTS', true);");
+            ResultSet rs = stat.executeQuery("SELECT * FROM TABLE_POINTS");
+            List<String> columns = JDBCUtilities.getColumnNames(rs.getMetaData());
+            assertEquals(2, columns.size());
+            assertTrue(columns.contains("THE_GEOM"));
+            assertTrue(columns.contains("ID"));
+            assertFalse(rs.next());
+        }
+    }
+
+    @Test
+    public void testWriteReadEmptyFGB2() throws Exception {
+        File file = new File("target/empty_file.fgb");
+        file.deleteOnExit();
+        try (Statement stat = connection.createStatement()) {
+            stat.execute("DROP TABLE IF EXISTS TABLE_POINTS");
+            stat.execute("create table TABLE_POINTS(id int, the_geom GEOMETRY(POINT, 4326))");
+            stat.execute("INSERT INTO TABLE_POINTS VALUES(1, 'SRID=4326;POINT(2 2)')");
+            stat.execute("CALL FGBWrite('target/empty_file.fgb', '(SELECT * FROM TABLE_POINTS WHERE the_geom && ST_BUFFER(''POINT(-1 -1)''::GEOMETRY, 0.001))', true);");
+            stat.execute("DROP TABLE IF EXISTS TABLE_POINTS");
+            stat.execute("CALL FGBRead('target/empty_file.fgb', 'TABLE_POINTS', true);");
             ResultSet rs = stat.executeQuery("SELECT * FROM TABLE_POINTS");
             List<String> columns = JDBCUtilities.getColumnNames(rs.getMetaData());
             assertEquals(2, columns.size());